MAINT: fix deprecated Python C API usage in odr
`PyEval_*` is deprecated although not being removed apparently.
Still, better to replace with the equivalent `PyObject_` call.
See https://bugs.python.org/issue42181

The printf cleanup is fixing an annoying warning; using a format
specifier like `%ld` or `%lld` that's hidden in a macro makes gcc
emit a warning that there are too many arguments (see
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=47781). So instead,
upcast the integer with unknown size to the largest possible size,
and use a fixed `%lld` specifier.