ENH: add callback to COBYLA optimizer
Addresses issue gh-2063 (currently closed, but unresolved). It adds the
callback function to the COBYLA optimizer in `minimize`.

I fully realize that in the COBYLA algorithm, each iteration is just a
single function evaluation, and that the user is free to do whatever
they want to in terms of callback behavior in that function. However,
currently the implementation of how the callback argument is handled is
just inconsistent between COBYLA and the other methods. Furthermore, a
user may want to separate their objective function from the callback.
This PR simply brings the COBYLA method in line with all the other
methods.