MAINT: silence warning in odepackmodule.c
The warning was:
```
../scipy/integrate/_odepackmodule.c: In function 'odepack_odeint':
../scipy/integrate/_odepackmodule.c:699:20: warning: 'tcrit' may be used uninitialized in this function [-Wmaybe-uninitialized]
  699 |         rwork[0] = *tcrit;
      |                    ^~~~~~
```

Also combines two if-clauses that did the same check but were weirdly
split, making the code harder to understand.

This fix is slightly obscure; it's not clear that subsequent calls
to LSODA inside the while-loop cannot set `itask=4`. And at that
point something will fail (it now will try using a NULL pointer
rather than an uninitialized one).

`itask=4` means that the ODE being solved has critical points.
Presumably that will be detected on the first call and not after
time-stepping, but that's hard to verify.