MAINT: _lib: add __dealloc__ to MessageStream
Extension class `MessageStream` doesn't implement a `__dealloc__` method which is required to free C level structures once the object goes out of scope. It is possible to do so with a call to the `close()` method but it is very inconvenient to call it every time one doesn't need the message stream anymore. Also, there could be a memory leak if one forgot to call the `close` method. Using `__dealloc__`, we can tell Cython to manage memory automatically for us.