BUG: Fixed various issues in scipy.sparse that lead to segfaults. csr_matrix can be constructed with out-of-bounds indices, leading to later segfaults #8778 Segfault when building a CSR sparse matrix #9253
Note that this has performance implications for loading sparse matrices using scipy.sparse.load_npz or when constructing sparse matrices using the tuple constructor. Invalid sparse matrices likely lead to a segfault, and are a security concern due to a "write-what-where" vulnerability in which an attacker could form a sparse matrix that reads/writes any address relative to the matrix object in memory. Because it uses the tuple version of the constructor, this can occur when load_npz is called to load a sparse matrix from a file.