ENH: Set fft backend with try_last=True (#14359)
#### Reference issue
This is a follow up to the #14266 #14275 

#### What does this implement/fix?
Now that the vendored uarray version is updated, this enables the functionality to use `try_last=True` when setting the global default backend in `scipy.fft`.

The following is an example snippet of what is now possible.
```python
import numpy as np
import cupy as cp
import scipy.fft
import cupyx.scipy.fft as cu_fft

# Register the cupy backend at the start
scipy.fft.register_backend(cu_fft)

# Basic
y_numpy = scipy.fft.fft(np.arange(10))

####### Earlier #######
# scipy with cupy array
x = cp.arange(10)
y_cupy = scipy.fft.fft(x)  # this will now raise a `TypeError` and complain about not using the correct backend

####### After the PR #######
# scipy with cupy array
x = cp.arange(10)
y_cupy = scipy.fft.fft(x) # works, no need of the context manager. just register your backend once at the top
```

#### How is this different from auto-dispatch proposed in #14266?
With auto-dispatch the proposal is to take this one step further by completely removing the backend registration at the top as well and placing that inside the other library (cupy for this example) which automatically registers on import. It is still open for discussion, irrespective of the decision to go ahead with auto-dispatch or not, this is still needed as part of the ***Step 2*** mentioned in #14266.