MAINT: bump scipy-mathjax
Same change as in https://github.com/numpy/numpy/pull/19314.
From that PR description:

This is actually a no-op (npm is used only if you upgrade the mathjax
stored in scipy-mathjax to a newer version manually, last done in 2017),
but silences some code quality scanners.

The package.json, rebuild.sh could also be removed from the shipped
documentation as they do nothing there, but not sure if that's worth the
hassle.

[ci skip]