MAINT: Begin translation of hyp2f1 for complex numbers into Cython (#14454)
This re-implements some special cases of the Gaussian
hypergeometric function for complex numbers into Cython, falling
back to the original Fortran implementation for unhandled
cases. It is the second in a series that also includes #14287,
which added extensive testing for hyp2f1 for complex
numbers. These two PRs contain portions of the work from #14256,
a closed PR which is in the process of being broken up into
smaller pieces that are less burdensome to review.