ENH: TST: add a threadpoolctl hook to limit OpenBLAS parallelism
This is taken over with minor modifications from scikit-learn.
The most dramatic effect of this is for `scipy.linalg`, its test
suite runs in ~30 sec on a 12-core machine without this change,
and in ~5 sec with it when using `linalg.test(parallel=12)`. The full
test suite is also sped up significantly.

Closes gh-14425