MAINT: Translation of hyp2f1 for complex numbers into Cython, part 2 (#14692)
This PR is the third in an ongoing project to translate the Gaussian hypergeometric function (exposed through `scipy.special.hyp2f1`) into Cython. To recap: the original Fortran implementation has many defects but the difficulty of working with and reviewing many lines of dense and sparsely documented Fortran 77 code has thwarted previous efforts to correct them. The goal is to move parts of the implementation in Cython piece by piece while correcting the defects until at last a solid and maintainable implementation is completed.

This PR handles the following cases

1. Unital argument for c - a - b > 0 through [DLMF 15.4.20](https://dlmf.nist.gov/15.4#E20).
2. z = -1 and c = 1 + a - b through [DLMF 15.4.26](https://dlmf.nist.gov/15.4#E26).
3. c - a or c - b a negative integer by applying an Euler hypegeometric transformation [DLMF 15.8.1](https://dlmf.nist.gov/15.8#E1) after which the computation reduces to evaluation of a polynomial.
4. Re(z) >= 0 and |z| < 0.9 through the Maclaurin series for hyp2f1.

For cases 1 and 2, where the function can be evaluated as a ratio of products of gamma functions, the behavior has been updated to compute with logs and then exponentiate if an underflow or overflow occurs with the standard method.

We've also updated the `max_degree` argument of `hyp2f1_series` to be a uint64. This gets rid of platform dependent behavior for the cases where computation reduces to polynomial evaluation.