MAINT: special: Updates for _cosine.c.
(1) In _cosine.c, change order of included headers.

polevl.h imports from numpy header files, which ultimately
include Python.h.  Python.h is supposed to be included before
other headers are included, so this changes moves the include
of polevl.h before that of <math.h>.

Also remove the inclusion of <stdio.h>, which is not used
in _cosine.c.

(2) Unconditionally define the macro M_PI64.

The code in the function cosine_cdf_pade_approx_at_neg_pi
requires a 64-bit approximation to pi.  So instead of
conditionally getting the definition from math.h, the macro
M_PI64 is defined in the file.