BUG: bsplines fixed issue 14525
bspline, quadratic and cubic gave zero outputs for integer dtypes.
This was caused by initialization of the output to the input dtype in cubic, quadratic, and np.piecewise (in bspline). In cubic and quadratic this was modified in place. For bspline it was fixed by casting the input to float (after abs to avoid issues with complex input).