CI: remove `pre_release_deps_source_dist` job from Azure CI config
See gh-14732 for context. The similarly named job
`prerelease_deps_64bit_blas` covers what we remove here. There are
other jobs that test with 32-bit BLAS on Linux. The job we leave runs
the full test suite, the one that is removed here the fast test suite.

This cuts down on CI time and also on duplicate failures.

[skip github]