correct the test programming error
One tests for correctness eigenpairs, in a loop,
which is meant to allow for a slack in Krylov subspaces method.

One should 1st check whether one got correct eigenvalues, and try again
if they are not the wanted ones.
The current implementation does it backwards: it checks whether
the eigenpair works, without 1st checking for eigenvalues.

This commit corrects this programming error