Add an array API to scipy.sparse
We add matching array classes for the bsr, coo, csc, csr, dia, dok,
and lil matrices.  For each of these, we restore element-wise
multiplication.

In the existing implementation, we switch to using the `@` operator
(instead of `*`, which behaves differently for arrays).

The existing matrix implementation often explicitly creates matrices
or sparse matrices.  To address that, we attach to each sparse array /
matrix various container attributes (`_coo_container`,
`_lil_container`, etc.).  These map to `_coo_matrix` and `_lil_matrix`
for sparse matrices, and to `_coo_array` and `_lil_array` for sparse
arrays.  Thus, the implementation can be agnostic of the container type.

For the same reason `_ascontainer` and `_container` methods are
introduced, which map to `asmatrix`/`matrix` and `asarray`/`array`
respectively.

* When slicing out a 1D result, return a dense array

We do this since there is no sparse mechanism for representing 1D arrays.

This commit also introduces `._is_array` that returns True for sparse
arrays but False for sparse matrices.

* Refactor sparray modifications into mixin

Also add __array__ to allow casting to ndarray

* Remove __array__ conversion

* Fix typo in division test

* Remove .A attribute

* Use _is_array throughout

* Remove .H in preference of .conj().T like in NumPy

* Change errors on .H and .A to VisibleDeprecationWarnings

* Move arrays to private _arrays