ENH: employ block structure to calculate Padé
The Padé problem is formulated and solved as

[B, A] [q, p]^T = 0

The matrix A, however has a very simple block structure:

A = [-I, 0]^T

where `I` is the identity matrix. We partition `B` likewise

B = [B_1, B_2]^T

and get the Block equation

[[B_1, -I], [[q],   [[0],
 [B_2,  0]]  [p]] =  [0]]

We can now solve the second line, which is a smaller matrix equation,
for `q` and substitute it in the first line for to obtain `p`.