BUG: stats: fix broadcasting behavior of argsreduce
`argsreduce` didn't broadcast the arrays when no invalid values were present in the given arguments. This led to wrong behaviour down the line for some methods of the distribution using it. This has been fixed and tests for one of the methods (`stats`) has been added.

I am trying to add tests that check broadcasting behavior for other methods that use `argsreduce`. This would a nice candidate for testing with Hypothesis.