TST: remove fragile test which checks if g77 is linked
We no longer need this test, because g77 is not a thing anymore
(the build will fail or other things will break if an ancient g77
is still used somehow).

This test is also fragile, as seen on conda-forge for the v1.7.2 builds,
the test will fail when the tests run under QEMU with:

E   RuntimeError: Failed to check dependencies for $PREFIX/lib/python3.9/site-packages/scipy/linalg/_flapack.cpython-39-aarch64-linux-gnu.so
        file       = '<...>/lib/python3.9/site-packages/scipy/linalg/_flapack.cpython-39-aarch64-linux-gnu.so'
        p          = <Popen: returncode: 1 args: ['ldd', '/home/conda/feedstock_root/build_artifa...>
        self       = <scipy.linalg.tests.test_build.FindDependenciesLdd object at 0x56782eb100>
        stderr     = b''
        stdout     = b'\tnot a dynamic executable\n'

That seems to be because `ldd` may incorrectly report something is not a
dynamic executable if the correct loader cannot be found.

[ci skip]