Resolve fft_interpolation_kind inconsistency
The fft pdf calculation uses scipy.interpolate.interp1d for interpolating,
whereas the fft cdf calculation uses scipy.interpolate's
InterpolatedUnivarateSpline. The latter has a method integral for
calculating the integral of the spline between two points which is
used in the cdf calculation. InterpolatedUnivariateSpline does not
have the kind parameter of interp1d which can take integer or string
arguments, but as a corresponding parameter k which only takes integer
arguments and specifies the degree of the spline. The existing
behavior is to not allow the spline degree to be specified for the fft
cdf calculation. This commit changes
both the pdf and cdf to use InterpolatedUnivariateSpline, and uses an
analog of fft_interpolation_kind in both cases. It is now called
fft_interpolation_degree to emphasize that it is not equivalent to the
kind parameter of interp1d. Separate class variables have also been
created for the pdf and cdf interpolation parameters.