Add reminder to restore faster indexing code path
Slicing that results in 1D sparse arrays is currently unsupported.
There was one such case in `linsolve.py`. We changed it to return a 2D
sparse array instead, but now also note that that is a slightly slower
code path, and that the old version should be restored once 1D sparse
arrays are supported.