ENH: added Fermi-Dirac distribution by name
The Fermi-Dirac distribution is heavily used in the field of
physics and it is obvious to retain this distribution function here.

It has been unclear to me whether the distribution should be named
differently since the actual Fermi-Dirac distribution as physicists
knows it is the survival function of the logistic distribution.