BUG: Fix hyp2f1 to return correct values in regions near exp(±iπ/3). (#14706)
This updates `scipy.special.hyp2f1` for complex `z` in the region
defined by 0.9 ≤ |z| < 1.1, |1 - z| ≥ 0.9, real(z) ≥ 0, to use
series expansion developed by José L. López and Nico M. Temme
which works well in this region. The existing implementation
struggles with points in this region. This method has been chosen
over Buhring's analytic continuation series which was used in
#8151 due to its simplicity in not requiring special handling
when `b - a` is (or is near) an integer, and because it appears
to perform better than Buhring's method in most cases.

`scipy.special._precompute.hyp2f1_data.py` has also been updated
to allow users the option to skip the mpmath calculations in
order to save time, since these calculations actually only need
to be performed once.