BUG: Deconvolve method does not work with signals with more than 1-D #14997
Closes #14997

What does it fix/implement?
The error described in the issue is caused by the method lfilter(),
which is called in the method deconvolve().
This method is made for 1-D arrays as it is defined in the docstrings.
It follows that the deconcolve() method can also be used only for
1-D input arrays (filter and divisor). The deconvolve() method in
_signaltools.py was therefore slightly modified and two
tests (test_n_dimensional_signal(), test_n_dimensional_divisor()) were
added to the corresponding test class TestDeconvolve in test_signaltools.py.
The fix consists of:
1) a change in the docstring method description so that the input parameter
is explicitly referred to as a 1-D array instead of an array.
2) a simple if statement that throws a value error when a multidimensional
array is specified
3) two additional tests to verify that the fix works.