BUG: special: Clean up some private namespaces and fix `special.__all__`
Before this change, `scipy.special.__all__` contained duplicated names.

From both `_ufuncs` and `_basic`:
    'gamma', 'hankel1', 'hankel2', 'hyp0f1', 'iv', 'jv', 'kv',
    'mathieu_a', 'mathieu_b', 'ndtri', 'psi', 'yv'

From both `_ufuncs` and `_orthogonal`:
    'binom', 'eval_chebyc', 'eval_chebys', 'eval_chebyt', 'eval_chebyu',
    'eval_gegenbauer', 'eval_genlaguerre', 'eval_hermite', 'eval_hermitenorm',
    'eval_jacobi', 'eval_laguerre', 'eval_legendre', 'eval_sh_chebyt',
    'eval_sh_chebyu', 'eval_sh_jacobi', 'eval_sh_legendre', 'poch'

The duplicated names are all ufuncs, so the true source of these names
is the `_ufuncs` module.  The new modules `_basic` and `_orthogonal`
are private, so there is no need to populate their namespaces with
names from `_ufuncs`.  I've cleaned up these modules so that their
`__all__` lists contain only the functions that are actually defined
there.

We *do* need to maintain the names in the deprecated public modules
`basic` and `orthogonal`, so I added the appropriate imports in
`basic.py` to maintain backwards compatibility.

I undid a couple obfuscatory aliases of `_ufuncs`; it was aliased
as `ufuncs` in `_basic.py` and as `cephes` in `_orthogonal.py`.

I also made a few incidental PEP 8 white space changes.