MAINT: Replace use of `pytest.warns(None)` with `warnings.catch_warnings`.
The use of `pytest.warns(None)` is deprecated in pytest 7.0.0.
The alternative suggested by the deprecation warning, `pytest.warns()`,
does not match the old behavior.  In fact, `pytest.warns()` is intended
specifically to raise an error if the code in the context does not raise
the given warning, and in pytest 7.0.0rc1, the default warning used by
`pytest.warns()` is the generic `Warning` class.

This issue is discussed in the following pytest issues:

* https://github.com/pytest-dev/pytest/issues/9002
* https://github.com/pytest-dev/pytest/issues/9386

There, the recommended alternative is to use `warnings.catch_warnings`.
That is the change made here.

Closes gh-15186