CI: remove coverage usage from Windows jobs
We seem to be measuring coverage on all jobs, and it remains quite
flaky. It is not necessary to do this on all jobs. So to stop the
hangs in CI, let's remove them from the Windows jobs where this is
a regular occurrence.

If the root cause is found, it can be re-enabled (preferably on only
a single Windows job with the full test suite, and not on all jobs -
we've had too many problems for too little gain for too long here).

Closes gh-14914