DOC: More docstring reformatting.
This try to autoformat the DocStrings a bit more, in particular focus on
the See Also sections, that for proper parsing are better with an
uppercase A. Also unlike in parameters sections when the dee also are a
definition list it is recommended to put a colon at the end of the line,
with space before colon. So this tries to do that, and when it can try
to fit it in one line (total length < 80 chars).

There is also some normalisations of having no spaces after headers,
and some normalisation of always having 4 spaces indent in rst.