Change default parameter for trace estimation
Trace estimation is quite costly, so for a single time step it is typically
hardly worth it. Thus we reduced it to `m3=1`, the correct order of magnitude
is sufficient.

For large number of time steps, the overhead of the trace estimation becomes
less relevant, and the algorithm might benefit more from an accurate estimate.
We choose `m3=5` which is quite arbitrary.