BLD: update NumPy to >=1.18.5, setuptools to <60.0
For NumPy, this is a regular bump after 1.8.x was branched (still
slightly on the conservative side w.r.t. NEP 29).

For setuptools, we do not want the vendored `distutils` that setuptools
re-enabled in 60.0. Given that we are dropping support for this build
system, it's not very likely we'll ever need >=60.0
This also fixes a CI failure we're seeing on 32-bit Linux:

  AttributeError: module '_distutils_hack' has no attribute 'ensure_shim'