TST: update `sparse.linalg` tests for failures due to tolerances (#15301)
* TST: bump test tolerance for propack float32/complex64 test

This test is new, and is showing failures on multiple configs
when testing 1.8.0rc2 on conda-forge. See
https://github.com/conda-forge/scipy-feedstock/pull/198#issuecomment-999180432

* TST: change seed for a lobpcg test

This test is new in 1.8.x, and was observed to fail on macOS + BLIS
in conda-forge testing. Should be a minor problem, but anyway having
a seed that's not zero is not a bad thing. So let's hope this is a more
robust seed.

* TST: bump tolerance on a sparse.linalg.lsqr test

Observed to slightly exceed the current tolerances on macOS + BLIS
in conda-forge testing of 1.8.0rc2 - see
https://github.com/conda-forge/scipy-feedstock/pull/198#issuecomment-999180432

* TST: bump test precision on a `sparse.linalg.svds` test

It was failing (only just) on macOS x86 + MKL in conda-forge testing:
```
___________ Test_SVDS_PROPACK.test_svd_random_state_3[random_state2] ___________
../[...]/lib/python3.8/site-packages/scipy/sparse/linalg/_eigen/tests/test_svds.py:368: in test_svd_random_state_3

    assert_allclose(uh_u, np.identity(k), atol=atol, rtol=rtol)
E   AssertionError:
E   Not equal to tolerance rtol=1e-06, atol=1e-10
E
E   Mismatched elements: 2 / 25 (8%)
E   Max absolute difference: 1.25789171e-10
E   Max relative difference: 4.21884749e-15
```

* TST: bump tolerance on `sparse.linalg.lsqr` test, add doc note