MAINT: Move z/(z - 1) transform case for hyp2f1 into Cython (#15287)
This PR is part of the ongoing work to rewrite scipy's Gaussian
hypergeometric function implementation for complex numbers in
Cython one small piece at a time. This PR handles the cases where
the z/(z - 1) transform (DLMF 15.8.1) is used. Where their
convergence regions overlap, the z/(z -1) transform is preferred
over the maclaurin series for real(z) < 0 because it avoids the
oscillation of terms in the maclaurin series for real(z) < 0 that
causes numerical issues due to cancellation.

The Fortran implementation in scipy/special/specfun/specfun.f
used this tranform when |z| < 1 and real(z) < 0, but the region
has been expanded to include points with |z| < 1.1. The Fortran
implementation uses the 1/z transform up until the very edge of
its region of convergence where it has numerical difficulties.

Comments have been updated to no longer refer cryptically to "The
original Fortran implementation" but instead to cite Shanjie
Zhang and Jianming Jin's implementation explictly.