CI: update Azure config for pyproject.toml defaulting to Meson
We need to use `python setup.py bdist_wheel` to build wheels with
distutils now (--no-pep-517 doesn't work, see discussion in gh-15476 for
details).

The CI job testing building from sdist (the one labelled
`source_distribution`) is switched over to Meson and the new
PyPI-specific sdist generation script.  This is an ubuntu-20 based job,
so a good test for meson-python (Ubuntu is a little "special").

Co-authored-by: Filipe Laíns <lains@riseup.net>