CI: remove matplotlib from 32-bit linux job, it fails to build
This was a pinned release, and building Matplotlib from source
was working 3 days ago but no longer does. It looks like a bug in
Pip 22.0 which was released 2 days ago.

We don't really need Matplotlib in this job, we only have a couple of
tests that actually need Matplotlib and those get skipped if it's not
installed. Building Matplotlib from source doesn't really make sense
here. So let's not do that - will speed up CI a bit as well.

Closes gh-15498

[skip github]