BUG: Avoid C Preprocessor symbol in _hypotests_pythran.py.
_P is a C Preprocessor symbol on my platform, so compilation of this file
fails with many errors, starting with syntax errors: expected identifier
before numeric constant and continuing with undeclared variables.

Renaming this function to have a longer name avoids this problem.
https://gcc.gnu.org/onlinedocs/gcc-11.2.0/cpp/System-specific-Predefined-Macros.html#System-specific-Predefined-Macros
suggests that symbols starting with an underscore and a capital letter
will perform unreliably when compiled, as they are part of the C
preprocessor reserved namespace.