DOC: update macOS and Linux contributor docs to use Python 3.9
This avoids an issue with M1. Python 3.8 had issues on M1 Macbooks,
so best to just use 3.9. We do release 3.8 wheels, but IIRC we needed
a fix in the wheels repo to make that work. M1 is too new to be
using 3.8 in development.

Also update the corresponding Linux instructions while we're at it.

Closes gh-15784

[ci skip]