DOC: Import Error in examples.
Trying to run the two examples fails.

I understand that this is in a private module. It was unclear to me if
those were meant to be public or not.
Alternative is to just remove the imports as those examples test the
functions themselves, and it it is common have  implied imports