DOC/MAINT: remove `quad_explain` that has become irrelevant.
As far as I can tell this was originally written by Travis in 2001,
before IPython's `?` and Python's help() would show docstring.

Through the years it has slowly mutate to do less and less until
it ended us just printing the docstring to stdout.

I don't believe this is necessary anymore.

This even leads to unnecessary churn, (like #7168, listing this function
as needing a docstring example).

In addition, the implementation is slightly incorrect: having
`sys.stdout` as a default value means it get evaluated at import time,
and this any context manger like contextlib's `redirect_stdout` will not
work as the reference is still to the original stdout. (This being the
main reason I found this while working on my new documentation project,
testing this functions broke the output capturing)

So all in all, I think removal is the right course.

[skip azp] [skip circle] [skip actions]