BUG: adds warning to scipy.stats.brunnermunzel
scipy.stats.brunnermunzel returns a p-value of nan
in some cases, because the numerator and denominator
of the t-distribution were 0. It is not clear
why nan is returned, especially when it appears
that p should be 0 (when effect sizes are large).
Now, a warning is emitted when p is nan, and an
alternative is suggested

Fixes #15843