BUG: special: Fix numerical precision issue of log_ndtr (#15187)
Reimplement `log_ndtr` to maintain consistently high precision.

The implementation is explained in the comments above the function
`faddeeva_log_ndtr()` in the file `_faddeeva.cxx`.

The test `TestNdtriExp.test_extreme` was rewritten. It tests the
round-trip calculation `x2 = log_ndtr(ndtri_exp(x))` at two extreme
input values.  The change of one of the extreme input values is
explained in the long comment in the test.

The other extreme value is changed to `-np.finfo(float).tiny`.  This
is because it made no sense to test that round-trip with the huge value
`-np.finfo(float).min`, which is the same as `np.finfo(float).max`,
approx. 1.8e308.  `ndtri_exp` returns `nan` for inputs greater than 0,
so the round-trip calculation returned `nan`.  The test passed because
the comparison function `assert_func_equal()` was used with `nan_ok=True`,
so the `nan` result was ignored (and so the test was pointless).  My
guess is the intent was to test the value that is now there, namely
`-np.finfo(float).tiny`, which is approximately -2.225e-308.  Because
this test is just a numerical comparison of two input values with two
output values, the simpler function `assert_allclose` is now used to
compare the results.

After updating `log_ndtr`, the stats test `TestExponNorm.test_std_pdf`
in `test_distributions.py` failed for two sets of parameter inputs,
(x, K) = (1, 0.01) and (x, K) = (-1, 0.01).  The calculation of
`exponnorm._logpdf` calls `log_ndtr(x - 1/K)` (via `_norm_logcdf`, which
is just a one-line wrapper of `log_ndtr`), and for these two sets of
parameters, the calculation of `log_ndtr(x - 1/K)` is *more* accurate
with this change than before.  So the increased error observed in the
test is the result of some other numerical effect.

Also, an 'Examples' section was added to the docstring of `log_ndtr`.

Closes gh-15172