ENH: add uarray multimethods for fast Hankel transforms (#15690)
I recently opened a CuPy PR adding GPU implementations of scipy.fft.fht and scipy.fft.ifht (cupy/cupy#6528), but noticed that we do not currently provide multimethods for these (they are not part of pocketfft and were added later in SciPy 1.7). This PR adds multimethods for these so that it is possible to provide them in 3rd party backends.