MAINT: revert change to `idzp_asvd0`
My review comment on the PR:

I looked at it again, this is definitely not correct. The first usage of
`list` in this routine is `call idz_copycols(m,n,a,krank,list,col)`;
`list` in that routine is also integer.

The problem with this code is deeper, `w(*)` is a work array declared as
`complex*16` but used for storing data of different types. Example on
line 83 of this file:
```
 call idzp_aid(eps,m,n,a,winit,krank,w(ilist),w(iproj))
```
The last two parameters are both `w`; the first one should be integer
and the second one complex.

In general this come is extremely poorly written. I think I'll revert
this change, merge the rest of the fixes, and then we can silence the
remaining warnings.