MAINT: Use lanczos approximation for gamma ratios in hyp2f1 for complex numbers (#15588)
This PR updates the cases z = 1, c - a - b > 0 [DLMF 15.4.20] and z near -1, c = 1 + a - b,
[DLMF 15.4.26] to use the Lanczos approximation to the gamma function for values that overflow or underflow with the direct computation of the gamma ratios, as suggested by @person142 in this comment on an earlier pull request in this series. This approach gives more accurate results than the naive approach of taking logs and then exponentiating at the end. This update has been made now because a function to compute these kinds of Gamma ratios using the Lanczos approximation will be helpful in the next two PR's which will implement the 1 - z [DLMF 15.8.4] and 1/z [DLMF 15.8.2] transforms respectively. These transforms are used to extend hyp2f1 outside the domain of its Maclaurin series.

In addition a bug has been fixed which has caused the cases where z is very near 1, c is a negative integer and one of a or b is a negative integer greater than or equal to c to be handled incorrectly. We seek to follow mpmath in returning finite values for these cases, using the result that in these cases the series becomes a terminating polynomial. These cases were being caught in earlier conditions before the terminating polynomial case was reached.