ENH/DEP: integrate: Use `opts` arg in dbl/tplquad
Allow the user more control over which integration method is used. This is
achieved by giving the user access to the `opts` parameters in `dblquad`
and `tplquad` with the default values of:

- epsabs = 1.49e-08
- epsrel = 1.49e-08
- limit  = 50
- points = None
- weight = None
- wvar   = None
- wopts  = None

This introduces some backwards compatability issues as `epsabs` and
`epsrel` are already passed as keyword arguments. For now these will
be ignored for now.