BUG: signal: Fix calculation of extended image indices in convolve2d.
To handle the `boundary` arguments "wrap" and "symm", the `convolve2d`
code computes indices into the first argument that might be negative or
larger than the size of the image.  The code adjusts negative or big
values so that the adjusted index refers to the correct data element
within the array.  The problem was that this adjustment code assumed
that it would only require that the data be extended *once*. E.g. if
an axis had length 100, it assumed that the extended index would be
in [-100, 200].  This is not true in general; the kernel (i.e. the
second argument of `convolve2d`) can be any size, and so the required
extension could be any size.  If the kernel was too big, the adjusted
indices might still be negative or too big, resulting in attempts to
access memory outside of input array's data.  That could result in
reading garbage, or a segmentation fault.

The fix is to use a couple helper functions that can handle arbitrarily
large kernel dimensions when computing the extended data.

Closes gh-8684.
Closes gh-8814.
Closes gh-12686.