BUG: Use relative error to check iircomb w0/fs
Checking for a remainder of zero when dividing floats is similar to
checking floats for exact equality; not a good practice.

For example, if fs = 100, it is possible to divide up the spectrum into
5 parts by specifying w0 = 20, but it is not possible to specify 11
parts using w0, because there is no floating point number that divides
evenly.

w0=9.090909090909090 produces N of 11.000000000000002 while
w0=9.090909090909091 produces N of 10.999999999999998

So we should instead check for approximate divisibility, with enough
precision to make it unlikely to be a coincidence.

I tested typical combinations of sample rates and integer divisors, and
found the relative error to be 2.2e-16 or less from one division
operation.