BUG: fix extension module initialization, needs use of `PyMODINIT_FUNC` (#16165)
This is related to gh-15996. After a change in Meson to use
visibility=hidden by default, imports were failing.

What is going on here is that CPython creates C/C++ default behavior of:
```
ifndef PyMODINIT_FUNC
       if defined(__cplusplus)
               define PyMODINIT_FUNC extern "C" Py_EXPORTED_SYMBOL PyObject*
       else /* __cplusplus */
               define PyMODINIT_FUNC Py_EXPORTED_SYMBOL PyObject*
       endif /* __cplusplus */
endif
```
and we, by using plain `PyObject *`, were missing the part where the
module initialization function was explicitly marked as exported.

It looks like we've never used `PyMODINIT_FUNC` correctly. The only
relevant change that I could find so quickly is gh-12086, which
removed `PyMODINIT_FUNC` from the only three modules that used it (and
so was incorrect).

Co-authored-by: Eli Schwartz <eschwartz@archlinux.org>