BUG: stats: Fix handling of float32 inputs for the boost-based ufuncs.
The C++14 code for calling the underlying C++ function that does
the actual scalar calculation for a ufunc had a bug where the
float arguments were implicitly cast to double while the function
expected float, resulting in incorrect output.

The fix here is to removed the C++14 version of the code, and
use only the less elegant C++11 code.

Closes gh-15961