Use real names when loading MATLAB opaque objects.
Loading MATLAB objects (i.e., instances of classdefs) from a .mat now
uses the objects actual names, instead of defaulting to the "None"
string, which led to overwriting issues when multiple objects are in the
same file.

whosmat used to raise a TypeError when loading such objects (due to a
`tuple(None)` call in shape_from_header), which prevented e.g. reading
the rest of the contents of the mat file.  As a partial fix, make it
return None instead for such objects.  The whosmat test also needs to be
fixed to get rid of the `__function_workspace__` metaentry that is added
for mat-files-with-objects (documented in _mio5.py).